GCP-Status-Kohls/GCP-Status-Kohls/requirements.txt
flask
google-api-python-client
google-cloud
pyOpenSSL==19.0.0
Flask-Login==0.4.1
oauthlib==2.0.6
requests==2.21.0
Flask-OAuth
gunicorn
oauth2client
datetime
pytz
requests-oauthlib==0.8.0
oauth2client
Authlib
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
firebase-admin
httplib2
Flask-OAuthlib
